Add files via upload
It's easy enough to just have a list of jobs one has applied to. Probably on scraps of paper, or memorized, or even, when one has the patience, to write it down in a journal to hopefully not forget about it. Luckily, with the amount of time I had not working due to the pandemic, I used that time to apply to jobs and create an entire spreadsheet down to color coding and creating an entire legend to remember what everything was meant to do.